Fix prefetch tests
